post like dislike

+++ database/migrations/2023_10_22_041734_create_post_tags_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('post_tag', function (Blueprint $table) {
             $table->uuid('post_id');
-            $table->unsignedInteger('tag_id');
+            $table->unsignedBigInteger('tag_id');
             $table->timestamps();
 
             $table->foreign('post_id')
